+++ src/parser/features/DDBFeatureMixin.js
@@ -990,7 +990,7 @@ export default class DDBFeatureMixin extends mixins.DDBActivityFactoryMixin {
     const activityData = foundry.utils.getProperty(this.data, `system.activities.${activity}`);
     if (activityData.type !== "summon") return activity;
     if (this.isCompanionFeature2014 || this.isCompanionFeature2024)
-      await this.ddbCompanionFactory.addCompanionsToDocuments([], activityData);
+      await this.ddbCompanionFactory.addCompanionsToDocuments([], activityData, this.enricher.activity);
     else if (this.isCRSummonFeature2024 || this.isCRSummonFeature2014)
       await this.ddbCompanionFactory.addCRSummoning(activityData);