+++ src/parser/companions/DDBSummonsManager.mjs
@@ -16,7 +16,8 @@ async function getSummonActors() {
   const dancingLights = DDBImporter.lib.DDBSummonsInterface.getDancingLights(jb2aMod);
   const mageHands = DDBImporter.lib.DDBSummonsInterface.getMageHands(jb2aMod);
   const bubblingCauldron = DDBImporter.lib.DDBSummonsInterface.getBubblingCauldrons();
-  const minorIllusions = DDBImporter.lib.DDBSummonsInterface.getMinorIllusions();
+  const illusions = DDBImporter.lib.DDBSummonsInterface.getIllusions();
+  const sensors = await DDBImporter.lib.DDBSummonsInterface.getClairvoyance();
   // const conjureAnimals = await getConjureAnimals();
 
   const localActors = {
@@ -24,7 +25,8 @@ async function getSummonActors() {
     ...dancingLights,
     ...mageHands,
     ...bubblingCauldron,
-    ...minorIllusions,
+    ...illusions,
+    ...sensors,
     // ...conjureAnimals,
   };